requirements.txt
flask==3.1.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.38.0
gunicorn==23.0.0
rapidfuzz==3.9.3
